requirements.txt
Django==3.1.7
psycopg2==2.8.6
gunicorn==20.0.4
djangorestframework
djangorestframework_jwt
djangorestframework-simplejwt